Fixed large files pull progress • Percentage is now calculated based on bytes transferred instead of using the provided percentage • Percentage is now displayed with 1 precision digit

+++ ADB Explorer/Models/File/SyncFile.cs
@@ -11,7 +11,7 @@ public class SyncFile : FilePath
 
     public FileOpProgressInfo LastUpdate => ProgressUpdates.LastOrDefault();
 
-    public int? CurrentPercentage => LastUpdate is AdbSyncProgressInfo adbInfo ? adbInfo.CurrentFilePercentage : null;
+    public double? CurrentPercentage => LastUpdate is AdbSyncProgressInfo adbInfo ? adbInfo.CurrentFilePercentage : null;
 
     public long? BytesTransferred => LastUpdate is AdbSyncProgressInfo adbInfo ? adbInfo.CurrentFileBytesTransferred : null;
